添加活动-列表查询 添加活动状态和停用

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/activity/ActivityListParam.java
@@ -14,13 +14,13 @@ import lombok.Setter;
 public class ActivityListParam extends BasePageParam {
 
     /** 进行中 **/
-    public static final int ONGOING = 0;
+    public static final byte ONGOING = 0;
     /** 未开始 **/
-    public static final int NOT_STARTED = 1;
+    public static final byte NOT_STARTED = 1;
     /** 已过期 **/
-    public static final int EXPIRED = 2;
+    public static final byte EXPIRED = 2;
     /** 已停用 **/
-    public static final int DISABLED = 3;
+    public static final byte DISABLED = 3;
 
     /** 状态 **/
     private Integer status;
